refactor: remove all genAi api files
