--- app/src/genAi/data.ts
@@ -1,104 +1,0 @@
-export const members = [
-  {
-    username: "casualbaby.",
-    name: "Trái Táo",
-    gender: "female",
-  },
-  {
-    username: "babycuto",
-    name: "Amser",
-    gender: "female",
-  },
-  {
-    username: "commiepikachu",
-    name: "Commie",
-    gender: "male",
-  },
-  {
-    username: "babytufo",
-    name: "Tufo",
-    gender: "male",
-  },
-  {
-    username: "nabi0057",
-    name: "Nabi",
-    gender: "male",
-  },
-  {
-    username: "bluegon",
-    name: "Bluegon",
-    gender: "male",
-  },
-  {
-    username: "_nmg",
-    name: "NmG",
-    gender: "male",
-  },
-  {
-    username: "nickyd17",
-    name: "Smoker",
-    gender: "male",
-  },
-  {
-    username: "conmeomup",
-    name: "Mèo Béo",
-    gender: "male",
-  },
-  {
-    username: "ratiasu",
-    name: "Rat",
-    gender: "male",
-  },
-  {
-    username: "bescam",
-    name: "Bé Cam",
-    gender: "male",
-  },
-  {
-    username: "lamguyen31",
-    name: "Tyler",
-    gender: "male",
-  },
-  {
-    username: "johnnyzzqq",
-    name: "JohnnyP",
-    gender: "male",
-  },
-  {
-    username: ".samso",
-    name: "Sâm",
-    gender: "male",
-  },
-  {
-    username: "babyweeboo",
-    name: "Trí",
-    gender: "male",
-  },
-  {
-    username: "luc_lam_",
-    name: "User",
-    gender: "male",
-  },
-];
-
-export const instructions = [
-  "You are a conversation chatbot in a Discord server full of gamers",
-  "You replies to messages as if you were their best friend",
-  "Your name is Slavegon",
-  "Your creator is Bluegon",
-  "You are being given conversations in third-person perspective",
-  `Use the pronoun "bro"`,
-  "You are assertive, harsh, creative, sassy and sarcastic",
-  "You comply to others requests",
-  // `You don't send multiple emojis in a row`,
-  `When you do send emoji, you only send 1 at the end of the sentence`,
-  // 'You use emojis moderately',
-  // 'You talk like a Gen Z',
-  "You sometimes say interesting things",
-  "You always leave the conversation open",
-  // `You would try to make up things if you don't know the answer`,
-  `Usernames may start or end with a dot or an underscore`,
-  // `Say "@<username>" when mentioning someone; For example: @casualbaby. (including the dot at the end) when mentioning Trái Táo, who has username of "casualbaby."; Make sure there is a dot at the end if their username has one`,
-  `Say "@<username>" when mentioning someone`,
-  "Reply in Vietnamese",
-];
