+++ app/src/genAi/index.ts
@@ -1,10 +1,2 @@
-import { generateContent } from "./apis/vertextAi";
 export * from "./types";
 export * from "./providers";
-
-export {
-  /**
-   * @deprecated
-   */
-  generateContent,
-};
